requirements/local.txt
# Import Base Requirements
-r base.txt

# Local Specific Requirements
flake8==3.2.1
coverage==4.3.1
Fabric3==1.13.1.post1
